+++ src/components/hptracker/HPTracker.tsx
@@ -1,16 +1,20 @@
-import React, { useEffect, useState } from "react";
+import React, { useEffect, useRef, useState } from "react";
 import { ContextWrapper } from "../ContextWrapper.tsx";
 import { usePlayerContext } from "../../context/PlayerContext.ts";
 import OBR, { Item } from "@owlbear-rodeo/sdk";
-import { characterMetadata } from "../../helper/variables.ts";
-import { HpTrackerMetadata } from "../../helper/types.ts";
-import { DragDropContext, Droppable, DropResult } from "react-beautiful-dnd";
+import { characterMetadata, sceneMetadata, version } from "../../helper/variables.ts";
+import { HpTrackerMetadata, SceneMetadata } from "../../helper/types.ts";
+import { DragDropContext, DraggableLocation, DropResult } from "react-beautiful-dnd";
 import "./hp-tracker.scss";
-import { DraggableTokenList, PlayerTokenList } from "./TokenList.tsx";
+import { PlayerTokenList } from "./TokenList.tsx";
 import { useCharSheet } from "../../context/CharacterContext.ts";
 import { CharacterSheet } from "./charactersheet/CharacterSheet.tsx";
 import { GlobalSettings } from "./globalsettings/GlobalSettings.tsx";
 import { SceneReadyContext } from "../../context/SceneReadyContext.ts";
+import { DropGroup } from "./DropGroup.tsx";
+import { sortItems } from "../../helper/helpers.ts";
+import changelog from "../../../CHANGELOG.md";
+import ReactMarkdown from "react-markdown";
 
 export const HPTracker = () => {
     return (
@@ -23,6 +27,11 @@ export const HPTracker = () => {
 const Content = () => {
     const playerContext = usePlayerContext();
     const [tokens, setTokens] = useState<Item[] | undefined>(undefined);
+    const [selectedTokens, setSelectedTokens] = useState<Array<string>>([]);
+    const [tokenLists, setTokenLists] = useState<Map<string, Array<Item>>>(new Map());
+    const [currentSceneMetadata, setCurrentSceneMetadata] = useState<SceneMetadata | null>(null);
+    const [changelogText, setChangelogText] = useState<string>("");
+    const dialogRef = useRef<HTMLDialogElement>(null);
     const { isReady } = SceneReadyContext();
     const { characterId } = useCharSheet();
 
@@ -34,6 +43,10 @@ const Content = () => {
                 (item.metadata[characterMetadata] as HpTrackerMetadata).hpTrackerActive
         );
         setTokens(initialItems);
+
+        const sceneData = await OBR.scene.getMetadata();
+        const metadata = sceneData[sceneMetadata] as SceneMetadata;
+        setCurrentSceneMetadata(metadata);
     };
 
     useEffect(() => {
@@ -52,24 +65,50 @@ const Content = () => {
             );
             setTokens(Array.from(filteredItems));
         });
+        OBR.scene.onMetadataChange((sceneData) => {
+            const metadata = sceneData[sceneMetadata] as SceneMetadata;
+            if (metadata) {
+                setCurrentSceneMetadata(metadata);
+            }
+        });
+        OBR.player.onChange((player) => {
+            setSelectedTokens(player.selection ?? []);
+        });
+        fetch(changelog)
+            .then((res) => res.text())
+            .then((text) => setChangelogText(text));
     }, []);
 
-    const sortItems = (a: Item, b: Item) => {
-        const aData = a.metadata[characterMetadata] as HpTrackerMetadata;
-        const bData = b.metadata[characterMetadata] as HpTrackerMetadata;
-        if (aData && bData && aData.index !== undefined && bData.index !== undefined) {
-            if (aData.index < bData.index) {
-                return -1;
-            } else if (aData.index > bData.index) {
-                return 1;
-            } else {
-                return 0;
-            }
-        }
-        return 0;
-    };
+    useEffect(() => {
+        const tokenMap = new Map<string, Array<Item>>();
 
-    const sortedTokens = Array.from(tokens ?? []).sort(sortItems);
+        currentSceneMetadata?.groups?.forEach((group) => {
+            const groupItems = tokens?.filter((item) => {
+                const metadata = item.metadata[characterMetadata] as HpTrackerMetadata;
+                return (
+                    (!metadata.group && group === "Default") ||
+                    metadata.group === group ||
+                    (!currentSceneMetadata?.groups?.includes(metadata.group ?? "") && group === "Default")
+                );
+            });
+            tokenMap.set(group, groupItems ?? []);
+        });
+
+        setTokenLists(tokenMap);
+    }, [currentSceneMetadata?.groups, tokens]);
+
+    const reorderMetadataIndex = (list: Array<Item>, group?: string) => {
+        OBR.scene.items.updateItems(list, (items) => {
+            items.forEach((item, index) => {
+                const data = item.metadata[characterMetadata] as HpTrackerMetadata;
+                data.index = index;
+                if (group) {
+                    data.group = group;
+                }
+                item.metadata[characterMetadata] = { ...data };
+            });
+        });
+    };
 
     const reorder = (list: Item[], startIndex: number, endIndex: number) => {
         const result = Array.from(list);
@@ -78,14 +117,23 @@ const Content = () => {
         result.splice(endIndex, 0, removed);
         const tokens = result.filter((item) => item !== undefined);
 
-        OBR.scene.items.updateItems(tokens, (items) => {
-            items.forEach((item, index) => {
-                const data = item.metadata[characterMetadata] as HpTrackerMetadata;
-                data.index = index;
+        reorderMetadataIndex(tokens);
+    };
 
-                item.metadata[characterMetadata] = { ...data };
-            });
-        });
+    const move = (
+        source: Array<Item>,
+        destination: Array<Item>,
+        droppableSource: DraggableLocation,
+        droppableDestination: DraggableLocation
+    ) => {
+        const sourceClone = Array.from(source);
+        const destClone = Array.from(destination);
+        const [removed] = sourceClone.splice(droppableSource.index, 1);
+
+        destClone.splice(droppableDestination.index, 0, removed);
+
+        reorderMetadataIndex(sourceClone);
+        reorderMetadataIndex(destClone, droppableDestination.droppableId);
     };
 
     const onDragEnd = (result: DropResult) => {
@@ -93,11 +141,21 @@ const Content = () => {
             return;
         }
 
+        if (result.source.droppableId != result.destination.droppableId) {
+            move(
+                tokenLists.get(result.source.droppableId) || [],
+                tokenLists.get(result.destination.droppableId) || [],
+                result.source,
+                result.destination
+            );
+            return;
+        }
+
         if (result.destination.index === result.source.index) {
             return;
         }
 
-        reorder(tokens ?? [], result.source.index, result.destination.index);
+        reorder(tokenLists.get(result.destination.droppableId) ?? [], result.source.index, result.destination.index);
     };
 
     return playerContext.role ? (
@@ -105,8 +163,21 @@ const Content = () => {
             <CharacterSheet />
         ) : (
             <div className={"hp-tracker"}>
-                <h1 className={"title"}>HP Tracker</h1>
-                {playerContext.role === "GM" ? <GlobalSettings /> : null}
+                <button className={"change-log-button"} onClick={() => dialogRef.current?.showModal()}>
+                    i
+                </button>
+                <dialog ref={dialogRef} className={"changelog"}>
+                    <button className={"change-log-button close"} onClick={() => dialogRef.current?.close()}>
+                        X
+                    </button>
+                    <ReactMarkdown children={changelogText} />
+                </dialog>
+                <h1 className={"title"}>
+                    HP Tracker<span className={"small"}>{version}</span>
+                </h1>
+                {playerContext.role === "GM" && !!currentSceneMetadata?.id ? (
+                    <GlobalSettings sceneId={currentSceneMetadata.id} />
+                ) : null}
                 <div className={`player-wrapper headings ${playerContext.role === "PLAYER" ? "player" : ""}`}>
                     <span>Name</span>
                     {playerContext.role === "GM" ? <span>Settings</span> : null}
@@ -115,19 +186,32 @@ const Content = () => {
                     <span className={"initiative-wrapper"}>INIT</span>
                     <span className={"character-sheet"}>INFO</span>
                 </div>
-                {playerContext.role === "GM" ? (
+                {playerContext.role === "GM" && currentSceneMetadata ? (
                     <DragDropContext onDragEnd={onDragEnd}>
-                        <Droppable droppableId={"tokens"}>
-                            {(provided) => (
-                                <div ref={provided.innerRef} {...provided.droppableProps}>
-                                    <DraggableTokenList tokens={sortedTokens} />
-                                    {provided.placeholder}
-                                </div>
-                            )}
-                        </Droppable>
+                        {currentSceneMetadata.groups && currentSceneMetadata.groups?.length > 0 ? (
+                            currentSceneMetadata.groups?.map((group) => {
+                                const list = tokenLists.get(group) || [];
+                                return (
+                                    <DropGroup
+                                        key={group}
+                                        title={group}
+                                        list={list.sort(sortItems)}
+                                        selected={selectedTokens}
+                                        metadata={currentSceneMetadata}
+                                    />
+                                );
+                            })
+                        ) : (
+                            <DropGroup
+                                title={"Default"}
+                                list={Array.from(tokens ?? []).sort(sortItems)}
+                                selected={selectedTokens}
+                                metadata={currentSceneMetadata}
+                            />
+                        )}
                     </DragDropContext>
                 ) : (
-                    <PlayerTokenList tokens={tokens ?? []} />
+                    <PlayerTokenList tokens={tokens ?? []} selected={selectedTokens} metadata={currentSceneMetadata!} />
                 )}
             </div>
         )